requirements.txt
flask
flask-cors
python-telegram-bot
nest-asyncio